fix : Replace myReviewThumb with myReviewWidget & modify file structure
